+++ tests/models/test_case_small_network_with_ates/src/run_ates.py
@@ -228,7 +228,7 @@ class HeatProblemPlacingOverTime(HeatProblem):
             )
             nominal = self.variable_nominal(f"{s}__cumulative_investments_made_in_eur")
             inv_cap = 2.5e5
-            constraints.append((inv_made[0] / nominal, 0.0, 200000.0))
+            constraints.append((inv_made[0] / nominal, 0.0, 200000.0 / nominal))
             for i in range(1, len(self.times())):
                 constraints.append(
                     (((inv_made[i] - inv_made[i - 1]) * nominal - inv_cap) / nominal, -np.inf, 0.0)
@@ -276,11 +276,8 @@ class HeatProblemSetPoints(
 
     def solver_options(self):
         options = super().solver_options()
-        options["solver"] = "highs"
-        highs_options = options["highs"] = {}
+        highs_options = options.setdefault("highs", {})
         highs_options["mip_rel_gap"] = 0.02
-        highs_options["presolve"] = "on"
-
         return options
 
     def constraints(self, ensemble_member):